Address latest review round: UsdSemantics as NoneType, prefixed World.create_with_root_body, move USD adapter into its own subpackage
- Use types.NoneType instead of None as the UsdSemantics fallback sentinel, so its
  type is preserved for readers.
- Change World.create_with_root_body to take a PrefixedName directly instead of a
  separate root_body_name/prefix pair of strings, updating every call site across
  the codebase (~50) to construct the PrefixedName explicitly.
- Move adapters/usd.py and adapters/usd_exceptions.py into a new
  adapters/usd/ subpackage (parser.py, exceptions.py), matching the
  artvip_dataset/procthor subpackage convention, and update every import.
- Confirmed empirically that the mjMESH_INERTIA_SHELL fix (mujoco 3.11) does not
  extend to MujocoBuilder._thicken_if_near_planar's coplanar-vertices workaround -
  that's a different MuJoCo convex-hull limitation, still reproduces on 3.11.0, so
  left unchanged.
- Documented in datasets.md that ArtVipDatasetLoader always loads a single named
  object (the catalog is one USD file per object), while USDParser itself can load
  a multi-object USD scene with no physics joints between the objects.

Made with the help of Claude.
