+++ semantic_digital_twin/scripts/new_robot_integration_helper.py
@@ -6,6 +6,7 @@ from semantic_digital_twin.adapters.ros.visualization.viz_marker import (
     VizMarkerPublisher,
 )
 from semantic_digital_twin.adapters.urdf import URDFParser
+from semantic_digital_twin.datastructures.prefixed_name import PrefixedName
 from semantic_digital_twin.robots.robot_parts import AbstractRobot
 from semantic_digital_twin.robots.unitree_g1 import UnitreeG1
 from semantic_digital_twin.world import World
@@ -19,7 +20,7 @@ robot_world = URDFParser.from_file(
 ).parse()
 robot_world.visualize_world_structure().show()
 
-world = World.create_with_root_body("map")
+world = World.create_with_root_body(PrefixedName("map"))
 world.merge_world(robot_world)
 
 rclpy.init()
