+++ semantic_digital_twin/src/semantic_digital_twin/adapters/usd/exceptions.py
