+++ semantic_digital_twin/src/semantic_digital_twin/predetermined_maps/apartment_environment.py
@@ -8,6 +8,7 @@ import numpy as np
 
 from semantic_digital_twin.adapters.package_resolver import CompositePathResolver
 from semantic_digital_twin.api import BodySpecification
+from semantic_digital_twin.datastructures.prefixed_name import PrefixedName
 from semantic_digital_twin.semantic_annotations.semantic_annotations import (
     Door,
     Handle,
@@ -47,7 +48,7 @@ class ApartmentEnvironment:
         """
         Create a world holding nothing but the apartment.
         """
-        world = World.create_with_root_body("root")
+        world = World.create_with_root_body(PrefixedName("root"))
         self.populate(world)
         return world
