+++ test/semantic_digital_twin_test/test_worlds/test_gcs_spawn_region.py
@@ -14,7 +14,7 @@ from semantic_digital_twin.spatial_types import HomogeneousTransformationMatrix
 
 @pytest.fixture
 def simple_world():
-    world = World.create_with_root_body("map")
+    world = World.create_with_root_body(PrefixedName("map"))
 
     # Add a target body
     target = Body(name=PrefixedName("target"))
